Refactoring: Bundling with Webpack, background ready for testing
