+++ docs/adding-webpack-notes.md
@@ -1,4 +1,7 @@
 ## References
 
+https://krzysztofzuraw.com/blog/2020/setting-up-chrome-extension-dev
 - https://github.com/sroze/ngInfiniteScroll#readme
 - https://stackoverflow.com/questions/38482692/cant-load-font-awesome-with-webpack
+- https://www.timroes.de/using-ecmascript-6-es6-with-angularjs-1-x
+- https://stackoverflow.com/questions/28969861/managing-jquery-plugin-dependency-in-webpack
