+++ src/ui/common.js
@@ -1,3 +1,5 @@
 function formatDate(date) {
 	return date.toLocaleDateString();
 }
+
+export {formatDate};
